:art: fix: 게스트하우스 삭제 에러 수정

See merge request s08-webmobile1-sub2/S08P12A603!458

Former-commit-id: 351d2d43b74ec866e03bc44fac06baf7385ba7f5

+++ backend/src/main/java/com/jejuinn/backend/api/controller/GuestHouseController.java
@@ -166,8 +166,9 @@ public class GuestHouseController {
         List<Recruitment> recruitment = recruitmentRepository.findAllByGuestHouseUidOrderByDateCreatedDesc(guestHouseUid);
         Long userUid = userService.getUserUidFromAccessToken(request);
         if(recruitment == null) return ResponseEntity.status(400).build();
-        System.out.println(recruitment.get(0).getGuestHouseUid());
-        recruitmentService.deleteRecruitment(recruitment.get(0).getUid(), userUid);
+        if(recruitment != null && recruitment.size() > 0) {
+            recruitmentService.deleteRecruitment(recruitment.get(0).getUid(), userUid);
+        }
         guestHouseRepository.deleteById(guestHouseUid);
         return ResponseEntity.status(200).build();
     }
